Adds i18n, which will be hooked up to a backend
The I18n backend will be hooked up to a database. Then the translations can be
loaded from there (cached, of course, locally based on timestamps), and the
frontend can have a client which can edit the copy _on the page_, and when it
gets saved, it'll be sent to the database (updating the changed translations,
and invalidating the cache).