Adds node-dev and removes the HTML files
node-dev lets the server restart on changes, as long as it's started with the
node-dev command (which I've added as `npm start`).

This commit also removes the HTML files from public, to check the dynamic
rendering from the markdown sources on the server. Similarly, the style.css
file is dynamically rendered now, too.