Fixes the fees table :sparkle:
The number of columns needs to be consistent, or the table looks a bit
janky. To this end, I moved the notice about the mandatory examination
into a table header (`thead`) with a single row `tr` containing just one
column (`td`) which spans 2 columns, matching the rest of the table.